Moved all database connectors to the Config.DatabaseConnectors namespace
