+++ mRemoteV1/Config/Connections/ConnectionsLoader.cs
@@ -1,4 +1,5 @@
 using mRemoteNG.App;
+using mRemoteNG.Config.DatabaseConnectors;
 using mRemoteNG.Config.DataProviders;
 using mRemoteNG.Config.Putty;
 using mRemoteNG.Config.Serializers;
