+++ mRemoteV1/Config/Connections/SqlConnectionsUpdateChecker.cs
@@ -4,6 +4,7 @@ using System;
 using System.Data;
 using System.Data.SqlClient;
 using System.Threading;
+using mRemoteNG.Config.DatabaseConnectors;
 
 namespace mRemoteNG.Config.Connections
 {
