+++ mRemoteV1/Config/DatabaseConnectors/IDatabaseConnector.cs
@@ -1,8 +1,7 @@
 ﻿using System;
 using System.Data.SqlClient;
 
-
-namespace mRemoteNG.Config
+namespace mRemoteNG.Config.DatabaseConnectors
 {
     public interface IDatabaseConnector : IDisposable
     {
